Fix out-of-range literals in tests
An 8-bit signed integer with a binary point at 2 effectively has a 6-bit
signed number for the value to the left of the decimal point. That means
it can represent values in the range [-2^5, 2^5-1], or [-32, 31]. After
https://github.com/chipsalliance/chisel/pull/4786 was merged, the `55`
and `56` literals are now correctly flagged as being out-of-range for
this type.

I've subtracted 32 from these out-of-range values (changing the MSB from
1 to 0), and the tests now pass.