Update Chisel from 6.5.0 to 7.0.0 (#11)
* Update Chisel from 6.5.0 to 7.0.0

`UnknownWidth` became a `case object` in this Chisel PR:
https://github.com/chipsalliance/chisel/pull/4242

The `connectFromBits` method was removed in this Chisel PR:
https://github.com/chipsalliance/chisel/pull/4168

The `connectFromBits` API was replaced with the `_fromUInt` API in this
Chisel PR: https://github.com/chipsalliance/chisel/pull/4782

Wrapping unary negation (`unary_-%`) was deprecated in this PR:
https://github.com/chipsalliance/chisel/pull/4829

The sbt update is necessary to maintain compatibility with the latest
Scala compiler version.

* Migrate from Chisel testers to ChiselSim

The `chisel3.testers` package was removed here:
https://github.com/chipsalliance/chisel/pull/4746

It doesn't look like ChiselSim has the same `Boolean` result from
`simulate`, so we must check for simulation errors by catching
exceptions.

* Replace ChiselRunners trait with direct use of ChiselSim

* Update to latest scalafmt

* lint: Apply latest scalafmt settings

* Update to latest scalatest version

* Fix missing sbt command in GitHub CI

For background: https://github.com/actions/setup-java/issues/712

Also updated JDK to 21 (latest LTS), updated some GitHub actions to
their latest versions, and using install-circt GitHub action to get
firtool.

* Fix out-of-range literals in tests

An 8-bit signed integer with a binary point at 2 effectively has a 6-bit
signed number for the value to the left of the decimal point. That means
it can represent values in the range [-2^5, 2^5-1], or [-32, 31]. After
https://github.com/chipsalliance/chisel/pull/4786 was merged, the `55`
and `56` literals are now correctly flagged as being out-of-range for
this type.

I've subtracted 32 from these out-of-range values (changing the MSB from
1 to 0), and the tests now pass.

* Update Chisel 7 snapshot from 639-5df5515f to 678-dca5fc11

This update includes this Chisel PR, which deprecated everything in the
`firrtl` package: https://github.com/chipsalliance/chisel/pull/4878

Bump `firtool` version to latest as well.

* Update Chisel version to 7.0.0-RC1

Also update CI versions of firtool and verilator to latest.

* Update Chisel version to 7.5.0